test(#1): move __tests__ to better location
