+++ pages/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import Home from '../pages/index';
+import Home from '../index';
 import '@testing-library/jest-dom';
 
 describe('Home', () => {
