+++ pages/__tests__/snapshot.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Home from '../pages/index';
+import Home from '../index';
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Home />);
